refactor : 오더 생성 로직 및 금액 계산 수정

+++ src/main/java/sparta/paymentassignment/domain/product/dto/ProductDetailResponse.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import sparta.paymentassignment.domain.product.entity.ProductStatus;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ProductDetailResponse {
     private final Long id;
     private final String name;
-    private final Long price;
+    private final BigDecimal price;
     private final int stock;
     private final String description;
     private final ProductStatus status;
