+++ src/main/java/sparta/paymentassignment/domain/product/dto/ProductSummaryResponse.java
@@ -3,11 +3,13 @@ package sparta.paymentassignment.domain.product.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class ProductSummaryResponse {
     private final Long id;
     private final String name;
-    private final Long price;
+    private final BigDecimal price;
     private final int stock;
 }
