+++ src/main/java/sparta/paymentassignment/exception/InvalidOrderAmountException.java
@@ -1,7 +1,9 @@
 package sparta.paymentassignment.exception;
 
+import java.math.BigDecimal;
+
 public class InvalidOrderAmountException extends RuntimeException {
-    public InvalidOrderAmountException(Long amount) {
+    public InvalidOrderAmountException(BigDecimal amount) {
         super("유효하지 않은 주문 금액입니다. amount=" + amount);
     }
 }
